refactor(remote): 收口远端准备包结构与命名
Why:
- 让 remote 主线对齐 unpack 与 mapping 的目录包结构
- 收掉旧单文件实现与冗余命名，降低后续维护成本

What:
- 新增 remote 包并拆分 preparer、lcu、game、cleanup 模块
- 更新 facade 与测试到新的 remote 包入口与最终命名
- 删除旧 remote_preparer 模块并补充 package API 回归测试

Risk:
- 兼容性影响：有，移除了旧模块路径与旧公开方法名
- 潜在风险：若仓库外仍有代码依赖旧 remote_preparer 导入，将需要同步迁移

Test:
- `uv run pytest -q tests/test_remote_package_api.py tests/test_remote_preparer.py tests/test_context_injection.py` -> `35 passed in 0.44s`
- `uv run ruff check src/lol_audio_unpack/remote src/lol_audio_unpack/facade.py tests/test_remote_package_api.py tests/test_remote_preparer.py tests/test_context_injection.py` -> `All checks passed!`
